requirements.txt
coveralls==1.1
dj-static==0.0.6
Django==1.9.6
django-envie==0.0.9
django-nose==1.4.3
djangorestframework==3.3.3
django-toolbelt==0.0.1
drfdocs==0.0.9
whitenoise==2.0.3
